start HashMap implementation
issue:
- ideally the structure will be
general purpose (can be used to
store data of multiple types
(especially because we want to be
able to use one of these for file
semaphores and one for the LRU
cache). However, not sure how to
make sure the void* data doesn't
get overwritten later on when new
semaphores, etc. are assigned to
the same variable that we pass
into the hashmap put function...